+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog of elfindo
 
-## v0.9 (work in progress)
+## v0.9.2 (work in progress)
+
+### New Features & Enhancements
+* New option "--copyto PATH" to copy all discovered regular files, symlinks, dirs from source to given destination.
+
+### Contributors
+* Thanks to Vang Le Quy for helpful comments and suggestions.
+
+## v0.9 (Dec 18, 2022)
 * Initial release by Sven Breuner.
 
 ### Contributors